feat: add build directory based install system.
feat: add pre/post install/remove functions.

A special directory is given for a package to build itself. And files
located in it will be moved to the root filesystem, as well as being
tracked by the package managers. Allowing the ability to get rid of
the remove function and just remove the associated files.